chore(tests): move ExtractNested to plugin_collection
