+++ falco_plugin_tests/src/plugin_collection/extract/mod.rs
@@ -1,3 +1,4 @@
 pub mod extra_fields;
+pub mod nested;
 pub mod remaining_from_table;
 pub mod remaining_from_table_runtime;
